add some platform specific code
(windows does not have syslog)
